requirements.txt
accelerate==0.24.0
albumentations==2.0.6
einops==0.8.1
flow_vis==0.1
h5py==3.13.0
huggingface_hub==0.31.1
hydra-core==1.3.2
imageio==2.37.0
ipython==8.12.3
joblib==1.5.0
kornia==0.8.1
matplotlib==3.10.8
mkl_service==2.4.0
numpy==2.3.5
omegaconf==2.3.0
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
packaging==25.0
pandas==2.3.3
parameterized==0.9.0
Pillow==12.0.0
psutil==7.0.0
pyceres==2.6
pytlsd==0.0.2
pytorch3d.egg==info
scikit_learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
setuptools==78.1.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.67.1
trimesh==4.10.1
visdom==0.2.4
